To reduce consumer friction, I've moved the iserializationstrategy classes out to the root to share same namespace as the consumer-facing classes, along with the queuesession interfaces. The remaining interfaces in the PublicInterfaces folder are not normally engaged by the consumer and have been moved back into an implementation namespace. This is a breaking change.
