+++ src/ModernDiskQueue/ISerializationStrategy.cs
@@ -1,4 +1,4 @@
-﻿namespace ModernDiskQueue.PublicInterfaces
+﻿namespace ModernDiskQueue
 {
     /// <summary>
     /// This class performs basic binary serialization from objects of Type T to byte arrays suitable for use in DiskQueue sessions.
