--- src/ModernDiskQueue/Implementation/ILockFile.cs
@@ -1,10 +1,0 @@
-using System;
-
-namespace ModernDiskQueue.Implementation
-{
-    /// <summary>
-    /// An inter-process lock based on a file.
-    /// The lock is removed by disposing of the instance.
-    /// </summary>
-    public interface ILockFile : IDisposable, IAsyncDisposable { }
-}
