+++ src/ModernDiskQueue/Implementation/PersistentQueueImpl.cs
@@ -1,8 +1,5 @@
 namespace ModernDiskQueue.Implementation
 {
-    using Microsoft.Extensions.Logging;
-    using Microsoft.Extensions.Logging.Abstractions;
-    using ModernDiskQueue.PublicInterfaces;
     using System;
     using System.Buffers;
     using System.Collections.Concurrent;
@@ -12,6 +9,10 @@ namespace ModernDiskQueue.Implementation
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
+    using Microsoft.Extensions.Logging;
+    using Microsoft.Extensions.Logging.Abstractions;
+    using ModernDiskQueue;
+    using ModernDiskQueue.Implementation.Interfaces;
 
     /// <summary>
     /// A persistent queue implementation.
