+++ src/ModernDiskQueue/ModernDiskQueue.csproj
@@ -54,6 +54,6 @@
     <TrimmerRootAssembly Include="Microsoft.Extensions.Logging.Abstractions" Version="8.0.2" />
     <TrimmerRootAssembly Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
     <TrimmerRootAssembly Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.2" />
-    <TrimmerRootAssembly Include="Microsoft.Extensions.Options" Version="8.0.2"/>
+    <TrimmerRootAssembly Include="Microsoft.Extensions.Options" Version="8.0.2" />
   </ItemGroup>
 </Project>
